+++ src/rules/pos-limit-rule.ts
